More performant main boot (saving 150ms)
